refactor: consolidate shared Firestore directory
