+++ src/shared/api.ts
@@ -1,0 +1,1 @@
+export { toRemoteById } from "./firestore";
