+++ src/shared/firebase.ts
@@ -1,0 +1,1 @@
+export { auth, db } from "./firestore/firebase";
