+++ src/shared/firestore/remoteSnapshot.ts
